2023-02-11 update(log activity)

+++ public/assets/js/customer_list.js
@@ -1,25 +1,30 @@
 console.log('st_customerList.js loading');
 $(document).ready(function () {
 
-    $('#myTable').DataTable({
-        responsive: true,
+    var table = $('#myTable').DataTable({
+        responsive: false,
         "order": [],
         "columns": [
             { "data": "thCustomerid" },
             { "data": "thName" },
-            { "data": "thAddress" },
-            { "data": "edit"},
-            { "data": "view"},
-            { "data": "delete"},
+            //{ "data": "thAddress" },
+            { "data": "edit" },
+            { "data": "view" },
+            { "data": "delete" },
         ],
         columnDefs: [
             { width: 100, targets: 0 },
-            { width: 30, targets: 3 },
-            { width: 30, targets: 4 },
-            { width: 30, targets: 5 }
+            //{ width: 200, targets: 2 },
+            { width: 50, targets: 2 },
+            { width: 50, targets: 3 },
+            { width: 50, targets: 4 },
 
-        ], 
+
+        ],
+
     });
+    //table.column( 4 ).visible( false );
+
     loadcustomers();
 })
 
@@ -39,7 +44,7 @@ function loadcustomers() {
                     data.push({
                         "thCustomerid": customer_id,
                         "thName": customer_name,
-                        "thAddress": address,
+                        //"thAddress": address,
                         "edit": '<button class="btn btn-primary" onclick="edit(' + customer_id + ')"><i class="fa fa-pencil-square-o" aria-hidden="true"></i></button>',
                         "view": '<button class="btn btn-success" onclick="view(' + customer_id + ')"><i class="fa fa-eye" aria-hidden="true"></i></button>',
                         "delete": '<button class="btn btn-danger" onclick="_delete(' + customer_id + ')"><i class="fa fa-trash" aria-hidden="true"></i></button>'
@@ -73,10 +78,14 @@ function loadcustomers() {
     });
 }
 function _delete(id) {
+    if (isAssignedCustomer(id)) {
+        showWarningMessage('This customer has already assigned to an order.')
+        return;
+    }
     $.ajax({
         type: 'DELETE',
         headers: { 'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content') },
-        url: '/st/CustomerList/deleteCustomer/' + id,
+        url: '/CustomerList/deleteCustomer/' + id,
         success: function (response) {
             console.log(response);
             loadcustomers();
@@ -86,6 +95,31 @@ function _delete(id) {
         }
     });
 }
+
+
+
+function isAssignedCustomer(id) {
+
+    var bool = 0;
+    $.ajax({
+        type: 'GET',
+        url: '/CustomerList/isAssignedCustomer/' + id,
+        async: false,
+        success: function (response) {
+            if (response == 1) {
+                bool = true;
+            } else {
+                bool = false;
+            }
+        }, error: function (data) {
+            // console.log('something went wrong');
+        }
+    });
+    return bool;
+}
+
+
+
 function edit(id) {
 
     location.href = "/customer?" + id;
